requirements.txt
beautifulsoup4
pytest-playwright
typing
flask
flask_sqlalchemy
flask_jwt_extended
flask_cors
pillow